Split off frame-handle from frame-element
