+++ ts/editor/FrameElement.svelte
@@ -3,11 +3,14 @@ Copyright: Ankitects Pty Ltd and contributors
 License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 -->
 <script lang="ts">
-    import { FrameStart, FrameEnd, FrameElement } from "../editable/frame-element";
+    import { FrameElement } from "../editable/frame-element";
+
+    customElements.define(FrameElement.tagName, FrameElement);
+
+    import { FrameStart, FrameEnd } from "../editable/frame-handle";
 
     customElements.define(FrameStart.tagName, FrameStart);
     customElements.define(FrameEnd.tagName, FrameEnd);
-    customElements.define(FrameElement.tagName, FrameElement);
 
     import { BLOCK_ELEMENTS } from "../lib/dom";
